remove bsb step from build pipeline
I was expecting to get rid of pm2 for watch at this point and only rely on webpack. Quite counterintuitively,
`webpack --watch` resulted in build times in excess of 2 seconds, while watching through pm2 and rerunning
webpack clocked in at around 500ms. This is pretty crazy and likely has to do with some kind of caching that
I'm not aware of right now.